[#132] Preliminary clean-up: moving global funcions from _init__.py to dedicated module
